Fix Errors

+++ allocators/src/character_array.c
@@ -6,7 +6,7 @@
 
 // Return the Length of the Dynamic Character Array
 int8_t array_length(char *array[]) {
-  int8_t index = 0;
+  int index = 0;
 
   while (*array[index] != '\0') {
     index = index + 1;
@@ -19,7 +19,7 @@ int8_t array_length(char *array[]) {
 // Assigns the Value of all Elements to '\0'
 // Returns the Array
 char clear_array(char *array[]) {
-  int8_t index;
+  int index;
   int8_t length = array_length(array);
 
   for (index = 0; index < length; index = index + 1) {
@@ -58,7 +58,7 @@ char create_array(char *array[], int8_t capacity) {
 // Shifts all Elements after it, to the Right
 // Returns the Array
 // Returns NULL if Memory Allocation Failed
-char insert_element(char *array[], char element, int8_t index) {
+char insert_element(char *array[], char element, int index) {
   int8_t address;
   int8_t length = array_length(array);
 
@@ -101,7 +101,7 @@ char push_element(char *array[], char element) {
 // Returns the Element
 // Shifts all Elements after it, to the Left
 // Returns NULL if Memory Allocation Failed
-char remove_element(char *array[], int8_t index) {
+char remove_element(char *array[], int index) {
   int8_t address;
   char *element = array[index];
   int8_t length = array_length(array);
