+++ allocators/src/integer_array.c
@@ -6,7 +6,7 @@
 
 // Return the Length of the Dynamic Integer Array
 int8_t array_length(int8_t *array[]) {
-  int8_t index = 0;
+  int index = 0;
 
   while (array[index] != NULL) {
     index = index + 1;
@@ -19,7 +19,7 @@ int8_t array_length(int8_t *array[]) {
 // Assigns the Value of all Elements to NULL
 // Returns the Array
 int8_t clear_array(int8_t *array[]) {
-  int8_t index;
+  int index;
   int8_t length = array_length(array);
 
   for (index = 0; index < length; index = index + 1) {
@@ -58,7 +58,7 @@ int8_t create_array(int8_t *array[], int8_t capacity) {
 // Shifts all Elements after it, to the Right
 // Returns the Array
 // Returns NULL if Memory Allocation Failed
-int8_t insert_element(int8_t *array[], int8_t element, int8_t index) {
+int8_t insert_element(int8_t *array[], int8_t element, int index) {
   int8_t address;
   int8_t length = array_length(array);
 
@@ -101,7 +101,7 @@ int8_t push_element(int8_t *array[], int8_t element) {
 // Returns the Element
 // Shifts all Elements after it, to the Left
 // Returns NULL if Memory Allocation Failed
-int8_t remove_element(int8_t *array[], int8_t index) {
+int8_t remove_element(int8_t *array[], int index) {
   int8_t address;
   int8_t *element = array[index];
   int8_t length = array_length(array);
