+++ allocators/src/integer_array.h
@@ -21,7 +21,7 @@ int8_t create_array(int8_t *array[], int8_t capacity);
 // Shifts all Elements after it, to the Right
 // Returns the Array
 // Returns NULL if Memory Allocation Failed
-int8_t insert_element(int8_t *array[], int8_t element, int8_t index);
+int8_t insert_element(int8_t *array[], int8_t element, int index);
 
 // Append an Element to the End of the Dynamic Integer Array
 // Returns the Array
@@ -33,6 +33,6 @@ int8_t push_element(int8_t *array[], int8_t element);
 // Returns the Element
 // Shifts all Elements after it, to the Left
 // Returns NULL if Memory Allocation Failed
-int8_t remove_element(int8_t *array[], int8_t index);
+int8_t remove_element(int8_t *array[], int index);
 
 #endif
